requirements.txt
pycrypto
paho-mqtt